Refactor: 폴더와 파일 구조 변경
